requirements.txt
dm_control>=1.0.20
gymnasium>=0.29.1
imageio>=2.34.0
mujoco>=3.1.6
numpy>=1.26.0
PyYAML>=6.0.1
shimmy[dm-control]>=1.3.0
stable-baselines3[extra]>=2.3.0
torch>=2.2.0
